+++ COFantasy/COFantasy.js
@@ -27,6 +27,7 @@ var COFantasy = COFantasy || function() {
 
   "use strict";
 
+  var PIX_PER_UNIT = 70;
   var DEF_MALUS_APRES_TOUR_5 = true;
   var IMAGE_OMBRE = "https://s3.amazonaws.com/files.d20.io/images/2781735/LcllgIHvqvu0HAbWdXZbJQ/thumb.png?13900368485";
   var IMAGE_DOUBLE = "https://s3.amazonaws.com/files.d20.io/images/33854984/q10B3KtWsCxcMczLo4BSUw/thumb.png?1496303265";
@@ -64,6 +65,7 @@ var COFantasy = COFantasy || function() {
     etourdi: 'status_half-haze',
     paralyse: 'status_fishing-net',
     ralenti: 'status_snail',
+    immobilise: 'status_cobweb',
     endormi: 'status_sleepy',
     apeure: 'status_screaming',
     invisible: 'status_ninja-mask',
@@ -311,6 +313,9 @@ var COFantasy = COFantasy || function() {
         removeFromTurnTracker(token.id, evt);
     }
     token.set(cof_states[etat], value);
+    if (!value) { //On enlève le save si il y en a un
+      removeTokenAttr(personnage, etat + 'Save', evt);
+    }
     if (etat == 'aveugle') {
       // We also change vision of the token
       if (aff.prev.light_losangle === undefined)
@@ -1125,6 +1130,27 @@ var COFantasy = COFantasy || function() {
     addCellInFramedDisplay(display, cell, 80, false);
   }
 
+  //options peut avoir les champs:
+  // - strict1 = true si on considère que tok1 doit avoir une taille nulle
+  // - strict2
+  // - allonge
+  function distanceCombat(tok1, tok2, pageId, options) {
+    if (pageId === undefined) {
+      pageId = tok1.get('pageid');
+    }
+    var page = getObj("page", pageId);
+    var scale = page.get('scale_number');
+    var pt1 = tokenCenter(tok1);
+    var pt2 = tokenCenter(tok2);
+    var distance_pix = VecMath.length(VecMath.vec(pt1, pt2));
+    options = options || {};
+    if (!options.strict1) distance_pix -= tokenSize(tok1, PIX_PER_UNIT);
+    if (!options.strict2) distance_pix -= tokenSize(tok2, PIX_PER_UNIT);
+    if (options.allonge) distance_pix -= (options.allonge * PIX_PER_UNIT) / scale;
+    if ((!options.strict1 || !options.strict2) && distance_pix < PIX_PER_UNIT * 1.5) return 0; //cases voisines
+    return ((distance_pix / PIX_PER_UNIT) * scale);
+  }
+
   // actif est un argument optionnel qui représente (token+charId) de l'unique token sélectionné
   function getSelected(msg, callback, actif) {
     var pageId = getPageId(msg);
@@ -1286,7 +1312,9 @@ var COFantasy = COFantasy || function() {
                 }
                 actif = tokenOfId(msg.selected[0]._id, msg.selected[0]._id, pageId);
               }
-              if (distanceCombat(centre.token, actif.token, pageId, true) > portee) {
+              if (distanceCombat(centre.token, actif.token, pageId, {
+                  strict1: true
+                }) > portee) {
                 sendChar(actif.charId, "Le centre de l'effet est placé trop loin (portée " + portee + ")");
                 return;
               }
@@ -1309,7 +1337,9 @@ var COFantasy = COFantasy || function() {
               if (obj.get('bar1_max') == 0) return; // jshint ignore:line
               var objChar = getObj('character', objCharId);
               if (objChar === undefined) return;
-              var distanceCentre = distanceCombat(centre.token, obj, pageId, true);
+              var distanceCentre = distanceCombat(centre.token, obj, pageId, {
+                strict1: true
+              });
               if (distanceCentre > rayon) return;
               selected.push({
                 _id: obj.id
@@ -1354,6 +1384,20 @@ var COFantasy = COFantasy || function() {
       else dest = origin.who;
     }
     sendChat('COF', '/w "' + dest + '" ' + msg);
+  }
+
+  function isCarac(x) {
+    switch (x) {
+      case 'FOR':
+      case 'DEX':
+      case 'CON':
+      case 'SAG':
+      case 'INT':
+      case 'CHA':
+        return true;
+      default:
+        return false;
+    }
   }
 
   function jet(msg) {
@@ -1427,7 +1471,7 @@ var COFantasy = COFantasy || function() {
         return;
       }
       var caracteristique = cmd[1];
-      if (isNotCarac(caracteristique)) {
+      if (!isCarac(caracteristique)) {
         error("Caracteristique '" + caracteristique + "' non reconnue (FOR, DEX, CON, INT, SAG, CHA).", cmd);
         return;
       }
@@ -1954,6 +1998,22 @@ var COFantasy = COFantasy || function() {
             etat: etat,
             condition: condition
           };
+          if (cmd[0] == 'etat' && cmd.length > 3) {
+            if (isCarac(cmd[2])) {
+              lastEtat.saveCarac = cmd[2];
+              var opposition = tokenOfId(cmd[3]);
+              if (opposition) {
+                lastEtat.saveDifficulte = cmd[3] + ' ' + opposition.token.get('name');
+              } else {
+                lastEtat.saveDifficulte = parseInt(cmd[3]);
+                if (isNaN(lastEtat.saveDifficulte)) {
+                  error("Difficulté du jet de sauvegarde incorrecte", cmd);
+                  lastEtat.saveCarac = undefined;
+                  lastEtat.saveDifficulte = undefined;
+                }
+              }
+            }
+          }
           scope.etats.push(lastEtat);
           return;
         case "psi": //deprecated
@@ -2355,6 +2415,19 @@ var COFantasy = COFantasy || function() {
           scope.messages = scope.messages || [];
           scope.messages.push(cmd.slice(1).join(' '));
           return;
+        case 'allonge':
+          if (cmd.length < 2) {
+            error("Il manque le message après --allonge", cmd);
+            return;
+          }
+          if (options.allonge !== undefined) {
+            log("Redéfinition de l'allong");
+          }
+          options.allonge = parseFloat(cmd[1]);
+          if (isNaN(options.allonge)) {
+            error("L'argument de --allonge n'est pas un nombre", cmd);
+          }
+          return;
         default:
           sendChat("COF", "Argument de !cof-attack '" + arg + "' non reconnu");
       }
@@ -2507,14 +2580,14 @@ var COFantasy = COFantasy || function() {
     var carac2;
     if (cmd[1].length == 3) {
       carac1 = cmd[1];
-      if (isNotCarac(cmd[1])) {
+      if (!isCarac(cmd[1])) {
         error("Le premier argument de save n'est pas une caractéristique", cmd);
         return;
       }
     } else if (cmd[1].length == 6) { //Choix parmis 2 caracs
       carac1 = cmd[1].substr(0, 3);
       carac2 = cmd[1].substr(3, 3);
-      if (isNotCarac(carac1) || isNotCarac(carac2)) {
+      if (!isCarac(carac1) || !isCarac(carac2)) {
         error("Le premier argument de save n'est pas une caractéristique", cmd);
         return;
       }
@@ -3757,7 +3830,10 @@ var COFantasy = COFantasy || function() {
         if (options.targetFx) {
           spawnFx(targetToken.get('left'), targetToken.get('top'), options.targetFx, pageId);
         }
-        var distanceTarget = distanceCombat(targetToken, attackingToken, pageId, true, true);
+        var distanceTarget = distanceCombat(targetToken, attackingToken, pageId, {
+          strict1: true,
+          strict2: true
+        });
         var pta = tokenCenter(attackingToken);
         var ptt = tokenCenter(targetToken);
         switch (options.aoe.type) {
@@ -3837,7 +3913,9 @@ var COFantasy = COFantasy || function() {
               };
               if (getState(cible, 'mort')) return; //pas de dégâts aux morts
               var distanceCentre =
-                distanceCombat(targetToken, obj, pageId, true);
+                distanceCombat(targetToken, obj, pageId, {
+                  strict1: true
+                });
               if (distanceCentre > options.aoe.rayon) return;
               var objChar = getObj('character', objCharId);
               if (objChar === undefined) return;
@@ -3941,7 +4019,7 @@ var COFantasy = COFantasy || function() {
         return;
       }
     }
-    if (options.avecd12 && estAffaibli(attaquant)) {
+    if (options.avecd12 && (estAffaibli(attaquant) || getState(attaquant, 'immobilise'))) {
       sendChar(attackingCharId, "ne peut pas utiliser cette capacité quand il est affaibli.");
       return;
     }
@@ -3962,8 +4040,11 @@ var COFantasy = COFantasy || function() {
     });
     if (cibles.length === 0) return;
     //Prise en compte de la distance
+    var optDistance = {};
+    if (options.contact) optDistance.allonge = options.allonge;
     cibles = cibles.filter(function(target) {
-      target.distance = distanceCombat(attackingToken, target.token, pageId);
+      target.distance =
+        distanceCombat(attackingToken, target.token, pageId, optDistance);
       if (options.intercepter || options.interposer) return true;
       if (target.distance > portee && target.esquiveFatale === undefined) {
         if (options.aoe || options.auto) return false; //distance stricte
@@ -4314,6 +4395,9 @@ var COFantasy = COFantasy || function() {
     if (estAffaibli(attaquant)) {
       dice = 12;
       explications.push("Attaquant affaibli => D12 au lieu de D20 en Attaque");
+    } else if (getState(attaquant, 'immobilise')) {
+      dice = 12;
+      explications.push("Attaquant aimmobilisé => D12 au lieu de D20 en Attaque");
     }
     if (options.avecd12) dice = 12;
     var nbDe = 1;
@@ -5153,6 +5237,9 @@ var COFantasy = COFantasy || function() {
               if (testCondition(ce.condition, attaquant, [target], d20roll)) {
                 setState(target, ce.etat, true, evt);
                 target.messages.push(target.tokName + " est " + ce.etat + eForFemale(target.charId) + " par l'attaque");
+                if (ce.saveCarac) {
+                  setTokenAttr(target, ce.etat + 'Save', ce.saveCarac, evt, undefined, ce.saveDifficulte);
+                }
               } else {
                 if (ce.condition.type == "moins") {
                   target.messages.push(
@@ -5331,6 +5418,9 @@ var COFantasy = COFantasy || function() {
                       function(reussite, rolltext) {
                         if (!reussite) {
                           setState(target, ce.etat, true, evt);
+                          if (ce.saveCarac) {
+                            setTokenAttr(target, ce.etat + 'Save', ce.saveCarac, evt, undefined, ce.saveDifficulte);
+                          }
                         }
                         saves--;
                         afterSaves();
@@ -6453,23 +6543,6 @@ var COFantasy = COFantasy || function() {
     return 0;
   }
 
-  //strict1 = true si on considère que tok1 doit avoir une taille nulle
-  function distanceCombat(tok1, tok2, pageId, strict1, strict2) {
-    if (pageId === undefined) {
-      pageId = tok1.get('pageid');
-    }
-    var PIX_PER_UNIT = 70;
-    var page = getObj("page", pageId);
-    var scale = page.get('scale_number');
-    var pt1 = tokenCenter(tok1);
-    var pt2 = tokenCenter(tok2);
-    var distance_pix = VecMath.length(VecMath.vec(pt1, pt2));
-    if (!strict1) distance_pix -= tokenSize(tok1, PIX_PER_UNIT);
-    if (!strict2) distance_pix -= tokenSize(tok2, PIX_PER_UNIT);
-    if ((!strict1 || !strict2) && distance_pix < PIX_PER_UNIT * 1.5) return 0; //cases voisines
-    return ((distance_pix / PIX_PER_UNIT) * scale);
-  }
-
 
   function malusDistance(perso1, tok2, distance, portee, pageId, explications, ignoreObstacles) {
     if (distance === 0) return 0;
@@ -6489,7 +6562,6 @@ var COFantasy = COFantasy || function() {
         layer: "objects"
       });
     var mObstacle = 0;
-    var PIX_PER_UNIT = 70;
     var pt1 = tokenCenter(tok1);
     var pt2 = tokenCenter(tok2);
     var distance_pix = VecMath.length(VecMath.vec(pt1, pt2));
@@ -8119,7 +8191,7 @@ var COFantasy = COFantasy || function() {
           var header = ligne.split(':');
           if (header.length > 1) {
             var c = header.shift().trim().toUpperCase();
-            if (isNotCarac(c)) return;
+            if (!isCarac(c)) return;
             carac = c;
             ligne = header.join(':').trim();
           }
@@ -8292,6 +8364,13 @@ var COFantasy = COFantasy || function() {
             }
           }
         });
+      }
+      for (var etat in cof_states) {
+        var saveEtat = boutonSaveState(perso, etat);
+        if (saveEtat) {
+          ligne += saveEtat + '<br />';
+          actionsAAfficher = true;
+        }
       }
       if (actionsAAfficher) {
         // on envoie la liste aux joueurs qui gèrent le personnage dont le token est lié
@@ -8752,8 +8831,14 @@ var COFantasy = COFantasy || function() {
       if (bufDef > 0)
         addLineToFramedDisplay(display, "Défense temporairement modifiée de " + bufDef);
       for (var etat in cof_states) {
-        if (getState(perso, etat))
-          addLineToFramedDisplay(display, etat + eForFemale(charId));
+        if (getState(perso, etat)) {
+          var etext = etat;
+          if (etext.endsWith('e')) etext = etext.substring(0, etext.length - 1) + 'é';
+          etext += eForFemale(charId);
+          var saveEtat = boutonSaveState(perso, etat);
+          if (saveEtat) etext += ", " + saveEtat;
+          addLineToFramedDisplay(display, etext);
+        }
       }
       if (charAttributeAsInt(charId, 'DEFARMUREON', 1) === 0) {
         addLineToFramedDisplay(display, "Ne porte pas son armure");
@@ -9063,7 +9148,8 @@ var COFantasy = COFantasy || function() {
 
   function deTest(personnage) {
     var dice = 20;
-    if (estAffaibli(personnage)) dice = 12;
+    if (estAffaibli(personnage) || getState(personnage, 'immobilise'))
+      dice = 12;
     return dice;
   }
 
@@ -9551,10 +9637,6 @@ var COFantasy = COFantasy || function() {
     return undefined;
   }
 
-  function isNotCarac(x) {
-    return (x != 'FOR' && x != 'DEX' && x != 'CON' && x != 'SAG' && x != 'INT' && x != 'CHA');
-  }
-
   function estElementaire(t) {
     if (t === undefined) return false;
     return (t == "feu" || t == "froid" || t == "acide" || t == "electrique");
@@ -9575,6 +9657,27 @@ var COFantasy = COFantasy || function() {
       error("Le premier argument de !cof-set-state n'est pas un état valide", cmd);
       return;
     }
+    var save;
+    if (isCarac(cmd[2])) {
+      if (cmd.length < 4) {
+        error("Il manque la difficulté du jet de sauvegarde.", cmd);
+        return;
+      }
+      valeur = true;
+      save = {
+        carac: cmd[2]
+      };
+      var opposition = tokenOfId(cmd[3]);
+      if (opposition) {
+        save.difficulte = cmd[3] + ' ' + opposition.token.get('name');
+      } else {
+        save.difficulte = parseInt(cmd[3]);
+        if (isNaN(save.difficulte)) {
+          error("Difficulté du jet de sauvegarde incorrecte", cmd);
+          return;
+        }
+      }
+    }
     var evt = {
       type: "set_state",
     };
@@ -9593,9 +9696,126 @@ var COFantasy = COFantasy || function() {
       }
       iterSelected(selected, function(perso) {
         setState(perso, etat, valeur, evt);
+        if (save) {
+          setTokenAttr(perso, etat + 'Save', save.carac, evt, undefined, save.difficulte);
+        }
       });
       addEvent(evt);
     });
+  }
+
+  function textOfSaveState(etat) {
+    switch (etat) {
+      case 'immobilise':
+        return "se libérer";
+      case 'aveugle':
+        return "retrouver la vue";
+      case 'etourdi':
+        return "reprendre ses esprits";
+      case 'assome':
+        return "reprendre conscience";
+      case 'renverse':
+        return "se relever";
+      case 'endormi':
+        return "se réveiller";
+      case 'apeure':
+        return "retrouver du courage";
+      default:
+        return "ne plus être " + etat;
+    }
+  }
+
+  function saveState(msg) {
+    var options = parseOptions(msg);
+    var cmd = options.cmd;
+    if (cmd === undefined || cmd.length < 4 ||
+      !_.has(cof_states, cmd[1]) || !isCarac(cmd[2])) {
+      error("Paramètres de !cof-save-state incorrects", cmd);
+      return;
+    }
+    var etat = cmd[1];
+    var carac = cmd[2];
+    var titre = 'Jet de ' + carac + ' pour ' + textOfSaveState(etat);
+    getSelected(msg, function(selected) {
+      if (selected.length === 0) {
+        error("Pas de token sélectionné", msg.content);
+        return;
+      }
+      var pageId = options.pageId;
+      if (pageId === undefined) {
+        iterSelected(selected, function(perso) {
+          if (pageId) return;
+          pageId = perso.token.get('pageid');
+        });
+      }
+      var opposant = tokenOfId(cmd[3], cmd[4], pageId);
+      if (opposant) {
+        iterSelected(selected, function(perso) {
+          if (!getState(perso, etat)) {
+            sendChar(perso.charId, "n'est pas " + etat + eForFemale(perso.charId));
+            return;
+          }
+          var evt = {
+            type: titre
+          };
+          var display = startFramedDisplay(msg.playerid, titre, perso, opposant);
+          var explications = [];
+          testOppose(perso, carac, opposant, carac, explications, evt,
+            function(resultat, crit) {
+              if (resultat == 2) {
+                explications.push(perso.token.get('name') + " est toujours " + etat + eForFemale(perso.charId));
+              } else {
+                setState(perso, etat, false, evt);
+                explications.push(perso.token.get('name') + " n'est plus " + etat + eForFemale(perso.charId));
+              }
+              explications.forEach(function(e) {
+                addLineToFramedDisplay(display, e);
+              });
+              addEvent(evt);
+              sendChat("", endFramedDisplay(display));
+            }); //fin test opposé (asynchrone)
+        }); //fin iterSelected du cas avec opposant
+      } else {
+        var seuil = parseInt(cmd[3]);
+        if (isNaN(seuil)) {
+          error("La difficulté n'est pas un nombre", cmd);
+          return;
+        }
+        iterSelected(selected, function(perso) {
+          if (!getState(perso, etat)) {
+            sendChar(perso.charId, "n'est pas " + etat + eForFemale(perso.charId));
+            return;
+          }
+          var evt = {
+            type: titre
+          };
+          testCaracteristique(perso, carac, seuil, {}, evt, function(res) {
+            sendChar(perso.charId, titre + " : " + res.texte);
+            if (res.reussite) {
+              setState(perso, etat, false, evt);
+              sendChar(perso.charId, res.texte + " &ge; " + seuil + ", " + perso.token.get('name') + " n'est plus " + etat + eForFemale(perso.charId));
+            } else {
+              sendChar(perso.charId, res.texte + " &lt; " + seuil + ", " + perso.token.get('name') + " est toujours " + etat + eForFemale(perso.charId));
+            }
+            addEvent(evt);
+          }); //fin test carac
+        }); //fin iterSelected du cas sans opposant
+      }
+    }); //fin getSelected
+  }
+
+  //Renvoie false si le personnage n'a pas d'attribut etatSave
+  function boutonSaveState(perso, etat) {
+    var attr = tokenAttribute(perso, etat + 'Save');
+    if (attr.length === 0) return false;
+    attr = attr[0];
+    var carac = attr.get('current');
+    if (!isCarac(carac)) {
+      log("Caractéristiques du save contre " + etat + " de " + perso.token.get('name') + " n'est pas une caractéristique " + carac);
+      return false;
+    }
+    var b = bouton("!cof-save-state " + etat + ' ' + carac + ' ' + attr.get('max'), "Jet", perso);
+    return b + " de " + carac + " pour " + textOfSaveState(etat);
   }
 
   function updateInit(token, evt) {
@@ -11751,7 +11971,6 @@ var COFantasy = COFantasy || function() {
         }
         sendChar(charId, "lance un sort de mur de force");
         if (options.image && sphere) {
-          var PIX_PER_UNIT = 70;
           var page = getObj("page", pageId);
           var scale = page.get('scale_number');
           var diametre = PIX_PER_UNIT * (6 / scale);
@@ -13947,6 +14166,9 @@ var COFantasy = COFantasy || function() {
     if (estAffaibli(guerrier)) {
       dice = 12;
       explications.push("Attaquant affaibli => D12 au lieu de D20 en Attaque");
+    } else if (getState(guerrier, 'imobilise')) {
+      dice = 12;
+      explications.push("Attaquant immobilisé => D12 au lieu de D20 en Attaque");
     }
     var toEvaluateAttack = attackExpression(guerrier, 1, dice, critGuerrier, armeGuerrier);
     sendChat('', toEvaluateAttack, function(resAttack) {
@@ -13994,6 +14216,9 @@ var COFantasy = COFantasy || function() {
       if (estAffaibli(cible)) {
         dice = 12;
         explications.push("Défenseur affaibli => D12 au lieu de D20 en Attaque");
+      } else if (getState(cible, 'immobilise')) {
+        dice = 12;
+        explications.push("Défenseur immobilisé => D12 au lieu de D20 en Attaque");
       }
       toEvaluateAttack = attackExpression(cible, 1, dice, critCible, armeCible);
       sendChat('', toEvaluateAttack, function(resAttack) {
@@ -14149,6 +14374,9 @@ var COFantasy = COFantasy || function() {
         return;
       case "!cof-set-state":
         interfaceSetState(msg);
+        return;
+      case "!cof-save-state":
+        saveState(msg);
         return;
       case "!cof-degainer":
         degainer(msg);
@@ -15609,7 +15837,7 @@ var COFantasy = COFantasy || function() {
       attrsSave.forEach(function(attr) {
         var attrName = attr.get('name');
         var carac = attr.get('current');
-        if (isNotCarac(carac)) {
+        if (!isCarac(carac)) {
           error("Save par tour " + attrName + " mal formé", carac);
           count--;
           if (count === 0) addEvent(evt);
@@ -15712,26 +15940,10 @@ var COFantasy = COFantasy || function() {
     }
   }
 
-  function estSurMonture(perso, pageId) {
-    var attr = tokenAttribute(perso, 'monteSur');
-    if (attr.length === 0) return false;
-    if (pageId === undefined) pageId = perso.token.get('pageid');
-    var monture = tokenOfId(attr[0].get('current'), attr[0].get('max'), pageId);
-    attr[0].remove();
-    if (monture === undefined) {
-      sendChar(perso.charId, "descend de sa monture");
-      return true;
-    }
-    sendChar(perso.charId, "descend de " + monture.token.get('name'));
-    removeTokenAttr(monture, 'estMontePar');
-    removeTokenAttr(monture, 'positionSurMonture');
-    return true;
-  }
-
   function nePeutPasBouger(perso) {
     if (attributeAsBool(perso, 'peutEtreDeplace')) return false;
     if (isActive(perso)) {
-      if (attributeAsBool(perso, 'immobilise')) return true;
+      if (getState(perso, 'immobilise')) return true;
       return false;
     }
     return true;
@@ -15758,7 +15970,8 @@ var COFantasy = COFantasy || function() {
     };
     var x = token.get('left');
     var y = token.get('top');
-    if (prev && (prev.left != x || prev.top != y) && nePeutPasBouger(perso)) {
+    var deplacement = prev && (prev.left != x || prev.top != y);
+    if (deplacement && nePeutPasBouger(perso)) {
       sendChar(charId, "ne peut pas se déplacer.");
       sendChat('COF', "/w GM " +
         '<a href="!cof-deplacer-token ' + x + ' ' + y + ' --target ' + token.id + '">Déplacer </a>' +
@@ -15767,20 +15980,32 @@ var COFantasy = COFantasy || function() {
       token.set('top', prev.top);
       return;
     }
-    var monture = {
-      token: token,
-      charId: charId
-    };
     var pageId = token.get('pageid');
-    if (estSurMonture(monture, pageId)) return;
-    var attr = tokenAttribute(monture, 'estMontePar');
+    //On regarde d'abord si perso est sur une monture
+    var attr = tokenAttribute(perso, 'monteSur');
+    if (attr.length > 0) {
+      if (deplacement) {
+        attr[0].remove();
+        var monture = tokenOfId(attr[0].get('current'), attr[0].get('max'), pageId);
+        if (monture === undefined) {
+          sendChar(charId, "descend de sa monture");
+          return;
+        }
+        sendChar(charId, "descend de " + monture.token.get('name'));
+        removeTokenAttr(monture, 'estMontePar');
+        removeTokenAttr(monture, 'positionSurMonture');
+      }
+      return;
+    }
+    //si non, perso est peut-être une monture
+    attr = tokenAttribute(perso, 'estMontePar');
     if (attr.length === 0) return;
     var cavalier = tokenOfId(attr[0].get('current'), attr[0].get('max'), pageId);
     if (cavalier === undefined) {
       attr[0].remove();
       return;
     }
-    var position = tokenAttribute(monture, 'positionSurMonture');
+    var position = tokenAttribute(perso, 'positionSurMonture');
     if (position.length > 0) {
       var dx = parseInt(position[0].get('current'));
       var dy = parseInt(position[0].get('max'));
@@ -15791,7 +16016,7 @@ var COFantasy = COFantasy || function() {
     }
     cavalier.token.set('left', x);
     cavalier.token.set('top', y);
-    cavalier.token.set('rotation', monture.token.get('rotation') + attributeAsInt(monture, 'directionSurMonture', 0));
+    cavalier.token.set('rotation', token.get('rotation') + attributeAsInt(perso, 'directionSurMonture', 0));
   }
 
   function deplacerToken(msg) {